Remove LS1043aBoardPkg since simulator platform is no longer supported
The LS1043aBoardPkg was last verified on the m0133 simulator model and
is untested (and hence may be broken) for newer simulator model
versions.

Remove the same for now. Later down the road, we can try fixing the same
and adding it back to the list of supported LS1043A UEFI platforms.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@freescale.com>