ShellPkg: Fix erroneous Status returned by ShellOpenFileByName()
In ShellOpenFileByName() the file is opened using
gEfiShellProtocol->OpenFileByName().
It is supposed that if this call returns an EFI_ERROR, the function
should return that error immediately. However, this return was
missing, and if UnicodeCollationProtocol has not been located by this
time, the Status gets overwritten with LocateProtocol() call result,
which eventually erroneously returns EFI_SUCCESS to the Shell.c, and
this leads to attempt to execute a non-existent startup script, which
fails, and which in turn leads to Shell being unloaded with "Invalid parameter"
error. This patch fixes the bug.

Signed-off-by: Meenakshi Aggarwal <meenakshi.aggarwal@freescale.com>