LS1043 : Make rcw for UEFI and U-boot common
Currently we are using a NOR memory map in which u-boot and uefi
boot loader bot are present and are at seprate locations.
BUT since we can use either bootloader at a time, we need to use
common memory for bootloader and its environment. made the changes
for same. This will also eliminate the need for separate rcw for
u-boot and uefi.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Pankaj Bansal<pankaj.bansal@nxp.com>