LS1043A: PCI NIC card connected to PCIE2 is not working
The pci devices' memory map is wrong. There
should be single memory space for all pcie devices attached
to one controller. the 32 bit aligned and 64 bit aligned
memory spaces should be allocated from this space. The address
returned to caller is the offset from the PCI controller base address.
when a memory read/write request from/to an address comes,
first the Base address of PCI controller is added to address and then
read/write from/to the address is performed.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>